# openapi/basedflare.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Basedflare API",
    "description": "API for Basedflare control panel",
    "version": "0.0.1"
  },
...
  "servers": [
    {
      "url": "https://cp.basedflare.com",
      "description": "Production"
    },
    {
      "url": "https://cp.basedflare.net",
      "description": "Backup"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Development (local server)"
    }
  ],
  "paths": {
...
    "/forms/stream": {
      "post": {
        "tags": [
          "Live Streaming"
        ],
        "summary": "Add a new streamer and return the stream key",
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appName": {
                    "type": "string",
                    "example": "jimmyjohn123",
                    "description": "Name of the streamer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "streamKey": {
                      "type": "string",
                      "example": "f6a4f78ce72cb20c36c76c5509a4b052"
                    },
                    "_id": {
                      "type": "string",
                      "example": "67a822f4518bc836547c9b1a"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "Conflict (appName already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Stream key with this name already exists"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error inserting key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to create new stream key"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/forms/stream/{id}": {
      "delete": {
        "tags": [
          "Live Streaming"
        ],
        "summary": "Delete stream key. Deleting concludes the stream for this key if live.",
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Stream key id",
            "schema": {
              "type": "string",
              "example": "67a822f4518bc836547c9b1a"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/yourdomain.com/your-admissions-webhook-url": {
      "post": {
        "tags": [
          "Live Streaming (Forwarded Webhooks)"
        ],
        "summary": "Receive stream admissions webhook events. NOTE: sent to your custom URL specified in the \"Create a stream webhook\" API.",
        "operationId": "admissionsWebhook",
        "parameters": [
          {
            "$ref": "#/components/headers/X-BF-Signature"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "Stream start": {
                  "summary": "Opening Event",
                  "value": {
                    "direction": "incoming",
                    "protocol": "RTMP",
                    "status": "opening",
                    "time": "2025-03-29T05:55:43.394+00:00",
                    "url": "rtmp://ome-origin-us-1.bfcdn.host:1935/app/6755124635874905051+tester?key=xxxxxxxxxxxxxxxxxxxxxxxxx"
                  }
                },
                "Stream stop": {
                  "summary": "Closing Event",
                  "value": {
                    "direction": "incoming",
                    "protocol": "RTMP",
                    "status": "closing",
                    "time": "2025-03-29T05:55:47.790+00:00",
                    "url": "rtmp://ome-origin-us-1.bfcdn.host:1935/app/6755124635874905051+tester?key=xxxxxxxxxxxxxxxxxxxxxx",
                    "new_url": "rtmp://ome-origin-us-1.bfcdn.host:1935/app/6755124635874905051+tester?key=xxxxxxxxxxxxxxxxxxxxxxx"
                  }
                }
              },
              "schema": {
                "type": "object",
                "properties": {
                  "request": {
                    "type": "object",
                    "properties": {
                      "direction": {
                        "type": "string",
                        "enum": [
                          "incoming"
                        ]
                      },
                      "protocol": {
                        "type": "string",
                        "enum": [
                          "RTMP"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "opening",
                          "closing"
                        ]
                      },
                      "time": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "url": {
                        "type": "string",
                        "format": "uri"
                      },
                      "new_url": {
                        "type": "string",
                        "format": "uri"
                      }
                    },
                    "required": [
                      "direction",
                      "protocol",
                      "status",
                      "time",
                      "url"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/yourdomain.com/your-alert-webhook-url": {
      "post": {
        "tags": [
          "Live Streaming (Forwarded Webhooks)"
        ],
        "summary": "Receive stream alert webhook events. NOTE: sent to your custom URL specified in the \"Create a stream webhook\" API.",
        "operationId": "alertWebhook",
        "parameters": [
          {
            "$ref": "#/components/headers/X-BF-Signature"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messages": {
                    "type": "array",
                    "items": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "OK"
                              ],
                              "example": "OK"
                            },
                            "description": {
                              "type": "string",
                              "example": "The current status is good"
                            }
                          }
                        },
                        {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INGRESS_HAS_BFRAME"
                              ],
                              "example": "INGRESS_HAS_BFRAME"
                            },
                            "description": {
                              "type": "string",
                              "example": "There are B-Frames in the ingress stream"
                            }
                          }
                        },
                        {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INGRESS_LONG_KEY_FRAME_INTERVAL"
                              ],
                              "example": "INGRESS_LONG_KEY_FRAME_INTERVAL"
                            },
                            "description": {
                              "type": "string",
                              "example": "The ingress stream's current keyframe interval (4.1 seconds) is too long. Please use a keyframe interval of 4.0 seconds or less"
                            }
                          }
                        },
                        {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INGRESS_BITRATE_HIGH"
                              ],
                              "example": "INGRESS_BITRATE_HIGH"
                            },
                            "description": {
                              "type": "string",
                              "example": "The ingress stream's current bitrate (5369946 bps) is higher than the configured bitrate (4000000 bps)"
                            }
                          }
                        }
                      ]
                    }
                  },
                  "sourceInfo": {
                    "type": "object",
                    "properties": {
                      "createdTime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-29T05:12:12.365+00:00"
                      },
                      "name": {
                        "type": "string",
                        "example": "6755124635874905051+tester"
                      },
                      "sourceType": {
                        "type": "string",
                        "example": "Rtmp"
                      },
                      "sourceUrl": {
                        "type": "string",
                        "example": "TCP://123.123.123.123:50130"
                      },
                      "tracks": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/StreamTrack"
                        }
                      }
                    }
                  },
                  "sourceUri": {
                    "type": "string",
                    "example": "#default#app/6755124635874905051+tester"
                  },
                  "type": {
                    "type": "string",
                    "example": "INGRESS"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "headers": {
      "X-BF-Signature": {
        "name": "X-Bf-Signature",
        "in": "header",
        "description": "A SHA256 HMAC signature of the payload (request body). This signature should be checked by your backend to verify the integrity and authenticity of the request.",
        "type": "string",
        "example": "71f453558de50865fb6feec836f4ed868542664afac639bc070304259418bc52"
      }
    },
    "schemas": {
      "StreamTrack": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 0
          },
          "name": {
            "type": "string",
            "example": "Video"
          },
          "type": {
            "type": "string",
            "example": "Video"
          },
          "video": {
            "type": "object",
            "properties": {
              "bitrate": {
                "type": "string",
                "example": "4500000"
              },
              "bitrateAvg": {
                "type": "string",
                "example": "3541368"
              },
              "bitrateConf": {
                "type": "string",
                "example": "4500000"
              },
              "bitrateLatest": {
                "type": "string",
                "example": "3749610"
              },
              "bypass": {
                "type": "boolean",
                "example": false
              },
              "codec": {
                "type": "string",
                "example": "H264"
              },
              "deltaFramesSinceLastKeyFrame": {
                "type": "integer",
                "example": 1
              },
              "framerate": {
                "type": "integer",
                "example": 30
              },
              "framerateAvg": {
                "type": "integer",
                "example": 30
              },
              "framerateConf": {
                "type": "integer",
                "example": 30
              },
              "framerateLatest": {
                "type": "number",
                "example": 29.97
              },
              "hasBframes": {
                "type": "boolean",
                "example": false
              },
              "height": {
                "type": "integer",
                "example": 720
              },
              "keyFrameInterval": {
                "type": "number",
                "example": 29.731
              },
              "keyFrameIntervalAvg": {
                "type": "number",
                "example": 29.731
              },
              "keyFrameIntervalConf": {
                "type": "integer",
                "example": 0
              },
              "keyFrameIntervalLatest": {
                "type": "integer",
                "example": 30
              },
              "width": {
                "type": "integer",
                "example": 1280
              }
            }
          },
          "audio": {
            "type": "object",
            "properties": {
              "bitrate": {
                "type": "string",
                "example": "128000"
              },
              "bitrateAvg": {
                "type": "string",
                "example": "130576"
              },
              "bitrateConf": {
                "type": "string",
                "example": "128000"
              },
              "bitrateLatest": {
                "type": "string",
                "example": "130584"
              },
              "bypass": {
                "type": "boolean",
                "example": false
              },
              "channel": {
                "type": "integer",
                "example": 2
              },
              "codec": {
                "type": "string",
                "example": "AAC"
              },
              "samplerate": {
                "type": "integer",
                "example": 48000
              }
            }
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid input"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token for API authentication. Use 'Bearer {token}'"
      }
    }
  }
}